# swagger_doc/openapi.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "SleekFlow TODOs API Coding Test",
    "version": "0.9.0"
  },
  "paths": {
...
    "/api/user/workspace/": {
      "get": {
        "tags": [
          "user"
        ],
        "summary": "Get All Workspaces",
        "description": "For the user with username, get all workspaces he or she has.",
        "operationId": "get_all_workspaces_api_user_workspace__get",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Username",
              "type": "string"
            },
            "name": "username",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseModel"
                },
                "example": {
                  "msg": "Get all workspaces joined by \"testing\" successfully.",
                  "data": [
                    {
                      "workspace_default_name": "testing",
                      "workspace_alias": null
                    }
                  ],
                  "error": null,
                  "error_msg": null
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized Error / Token Expired Error / Invalid Token Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseModel"
                },
                "examples": {
                  "unauthorized_error": {
                    "value": {
                      "msg": null,
                      "data": null,
                      "error": "UnauthorizedError",
                      "error_msg": "Unauthorized action."
                    }
                  },
                  "token_expired_error": {
                    "value": {
                      "msg": null,
                      "data": null,
                      "error": "TokenExpiredError",
                      "error_msg": "Token has expired."
                    }
                  },
                  "invalid_token_error": {
                    "value": {
                      "msg": null,
                      "data": null,
                      "error": "InvalidTokenError",
                      "error_msg": "Invalid token."
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseModel"
                },
                "example": {
                  "msg": null,
                  "data": null,
                  "error": "ValidationError",
                  "error_msg": "Some Error Message"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseModel"
                },
                "example": {
                  "msg": null,
                  "data": null,
                  "error": "InternalServerError",
                  "error_msg": "Unknown Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/workspace/todolists/todos/": {
      "get": {
        "tags": [
          "workspace"
        ],
        "summary": "Get All Todolists Todos",
        "description": "Get a list of all workspaces",
        "operationId": "get_all_todolists_todos_api_workspace_todolists_todos__get",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Username",
              "type": "string"
            },
            "name": "username",
            "in": "query"
          },
          {
            "required": true,
            "schema": {
              "title": "Workspace Default Name",
              "type": "string"
            },
            "name": "workspace_default_name",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseModel"
                },
                "example": {
                  "msg": "Get all todolists and corresponding todos in workspace \"workspace_testing\" successfully.",
                  "data": [
                    {
                      "todolist_id": 1,
                      "todolist_name": "todolist_testing",
                      "todos": [
                        {
                          "todo_id": 1,
                          "todo_name": "testing",
                          "todo_description": "testing",
                          "todo_due_date": "2010-01-01T00:00:00.000Z",
                          "todo_priority": "low",
                          "todo_status": "None",
                          "todo_last_modified": "2022-10-21T13:42:21.000Z"
                        },
                        {
                          "todo_id": 2,
                          "todo_name": "testing2",
                          "todo_description": "new_tesing",
                          "todo_due_date": "2010-01-01T00:00:01.000Z",
                          "todo_priority": "high",
                          "todo_status": "finished",
                          "todo_last_modified": "2022-10-21T13:41:21.000Z"
                        }
                      ]
                    }
                  ],
                  "error": null,
                  "error_msg": null
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized Error / Token Expired Error / Invalid Token Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseModel"
                },
                "examples": {
                  "unauthorized_error": {
                    "value": {
                      "msg": null,
                      "data": null,
                      "error": "UnauthorizedError",
                      "error_msg": "Unauthorized action."
                    }
                  },
                  "token_expired_error": {
                    "value": {
                      "msg": null,
                      "data": null,
                      "error": "TokenExpiredError",
                      "error_msg": "Token has expired."
                    }
                  },
                  "invalid_token_error": {
                    "value": {
                      "msg": null,
                      "data": null,
                      "error": "InvalidTokenError",
                      "error_msg": "Invalid token."
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found Error, on User / Workspace",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseModel"
                },
                "examples": {
                  "user_not_found": {
                    "value": {
                      "msg": null,
                      "data": null,
                      "error": "NotFoundError",
                      "error_msg": "User \"testing\" not found."
                    }
                  },
                  "workspace_not_found": {
                    "value": {
                      "msg": null,
                      "data": null,
                      "error": "NotFoundError",
                      "error_msg": "Workspace \"workspace_testing\" not found."
                    }
                  },
                  "user_not_in_workspace": {
                    "value": {
                      "msg": null,
                      "data": null,
                      "error": "NotFoundError",
                      "error_msg": "User \"testing\" has not joined workspace \"workspace_testing\"."
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseModel"
                },
                "example": {
                  "msg": null,
                  "data": null,
                  "error": "ValidationError",
                  "error_msg": "Some Error Message"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseModel"
                },
                "example": {
                  "msg": null,
                  "data": null,
                  "error": "InternalServerError",
                  "error_msg": "Unknown Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/workspace/invite/": {
      "put": {
        "tags": [
          "workspace"
        ],
        "summary": "Invite User To Workspace",
        "description": "Invite user to a workspace",
        "operationId": "invite_user_to_workspace_api_workspace_invite__put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InviteWorkspaceModel"
              },
              "example": {
                "owner_username": "testing1,",
                "workspace_default_name": "workspace_testing,",
                "invitee_username": "testing2,"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseModel"
                },
                "example": {
                  "msg": "Invited user \"testing2\" to workspace \"workspace_testing\" successfully",
                  "data": null,
                  "error": null,
                  "error_msg": null
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized Error / Token Expired Error / Invalid Token Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseModel"
                },
                "examples": {
                  "unauthorized_error": {
                    "value": {
                      "msg": null,
                      "data": null,
                      "error": "UnauthorizedError",
                      "error_msg": "Unauthorized action."
                    }
                  },
                  "token_expired_error": {
                    "value": {
                      "msg": null,
                      "data": null,
                      "error": "TokenExpiredError",
                      "error_msg": "Token has expired."
                    }
                  },
                  "invalid_token_error": {
                    "value": {
                      "msg": null,
                      "data": null,
                      "error": "InvalidTokenError",
                      "error_msg": "Invalid token."
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found Error, on User / Workspace",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseModel"
                },
                "examples": {
                  "user_not_found": {
                    "value": {
                      "msg": null,
                      "data": null,
                      "error": "NotFoundError",
                      "error_msg": "User \"testing\" not found."
                    }
                  },
                  "workspace_not_found": {
                    "value": {
                      "msg": null,
                      "data": null,
                      "error": "NotFoundError",
                      "error_msg": "Workspace \"workspace_testing\" not found."
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "Duplicate Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseModel"
                },
                "example": {
                  "msg": null,
                  "data": null,
                  "error": "NotFoundError",
                  "error_msg": "User \"testing2\" has already joined workspace \"workspace_testing\""
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseModel"
                },
                "example": {
                  "msg": null,
                  "data": null,
                  "error": "ValidationError",
                  "error_msg": "Some Error Message"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseModel"
                },
                "example": {
                  "msg": null,
                  "data": null,
                  "error": "InternalServerError",
                  "error_msg": "Unknown Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "InviteWorkspaceModel": {
        "title": "InviteWorkspaceModel",
        "required": [
          "owner_username",
          "workspace_default_name",
          "invitee_username"
        ],
        "type": "object",
        "properties": {
          "owner_username": {
            "title": "Owner Username",
            "type": "string"
          },
          "workspace_default_name": {
            "title": "Workspace Default Name",
            "type": "string"
          },
          "invitee_username": {
            "title": "Invitee Username",
            "type": "string"
          }
        },
        "description": "Create workspace model class for api"
      },
...
      "ResponseModel": {
        "title": "ResponseModel",
        "required": [
          "msg",
          "data",
          "error",
          "error_msg"
        ],
        "type": "object",
        "properties": {
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "data": {
            "title": "Return data",
            "type": "string"
          },
          "error": {
            "title": "Error Type",
            "type": "string"
          },
          "error_msg": {
            "title": "Error Message",
            "type": "string"
          }
        },
        "description": "Response model"
      },
...
    }
  }
}